# auth-service/docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Auth Service API with PingAM Integration
  description: |
    Comprehensive authentication and authorization service with PingAM integration.
    
    **Features:**
    - User authentication and registration
    - PingAM SSO integration
    - OAuth2/OIDC support
    - Role-based access control (RBAC)
    - Permission management
    - Session management
    - Token management (access & refresh tokens)
    
    **Base URL:** `http://localhost:8085/api/v1`
  version: 1.0.0
  contact:
    name: Auth Service Team
    email: auth-service@example.com
servers:
  - url: http://localhost:8085/api/v1
    description: Local Development Server
  - url: http://localhost:1080
    description: PingAM Mock Server
  - url: http://localhost:8090
    description: PingAM Server (OpenAM)
...
paths:
...
  /auth/refresh-token:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get a new access token using refresh token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # ========== PingAM Authentication Endpoints ==========
...
  /users:
    post:
      tags:
        - User Management
      summary: Create user
      description: Create a new user
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '409':
          $ref: '#/components/responses/ConflictError'

    get:
      tags:
        - User Management
      summary: List users
      description: Get list of users with pagination
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Items per page
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  schemas:
...
    RefreshTokenRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
          example: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
...
    CreateUserRequest:
      type: object
      required:
        - username
        - email
        - password
      properties:
        username:
          type: string
          example: "newuser"
        email:
          type: string
          format: email
          example: "newuser@example.com"
        password:
          type: string
          format: password
          example: "SecurePass123!"

    # ===== Response Schemas =====
...
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: "Bearer"
        expires_in:
          type: integer
          example: 3600
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: string
        code:
          type: string

  # ===== Reusable Responses =====
  responses:
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Invalid request"
            details: "Missing required field: username"
            code: "BAD_REQUEST"
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Unauthorized"
            details: "Invalid or expired token"
            code: "UNAUTHORIZED"
...
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "Conflict"
            details: "User already exists"
            code: "CONFLICT"